refactor: First Class Collections
Wrap objects and methods in a class
